--- accumulo/src/main/resources/META-INF/services/geotrellis.layers.io.AttributeStoreProvider
@@ -1,1 +1,0 @@
-geotrellis.spark.io.accumulo.AccumuloLayerProvider
