+++ accumulo/src/main/resources/META-INF/services/geotrellis.spark.io.LayerWriterProvider
@@ -1,1 +1,1 @@
-geotrellis.spark.io.accumulo.AccumuloLayerProvider
+geotrellis.spark.io.accumulo.AccumuloSparkLayerProvider
