+++ accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerCopier.scala
@@ -16,17 +16,18 @@
 
 package geotrellis.spark.io.accumulo
 
-import geotrellis.layers.LayerId
 import geotrellis.tiling.{Boundable, Bounds, KeyBounds}
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.spark.io.AttributeStore.Fields
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.index._
-import geotrellis.layers.io.json._
+import geotrellis.layers._
+import geotrellis.layers.AttributeStore.Fields
+import geotrellis.layers.accumulo._
+import geotrellis.layers.avro._
+import geotrellis.layers.index._
+import geotrellis.layers.json._
 import geotrellis.util._
+
 import org.apache.avro._
 import org.apache.spark.SparkContext
+
 import org.apache.spark.rdd.RDD
 import spray.json.JsonFormat
