+++ accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerReindexer.scala
@@ -17,17 +17,19 @@
 package geotrellis.spark.io.accumulo
 
 import geotrellis.tiling.{Boundable, Bounds}
-import geotrellis.spark._
+import geotrellis.layers._
+import geotrellis.layers.accumulo._
+import geotrellis.layers.avro._
+import geotrellis.layers.index._
+import geotrellis.layers.json._
 import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
-import geotrellis.layers.io.index._
-import geotrellis.layers.io.json._
 import geotrellis.util._
-import org.apache.spark.SparkContext
-import spray.json.JsonFormat
-import java.time.ZonedDateTime
 
-import geotrellis.layers.LayerId
+import org.apache.spark.SparkContext
+
+import spray.json.JsonFormat
+
+import java.time.ZonedDateTime
 
 import scala.reflect.ClassTag
