+++ layers-accumulo/src/main/resources/META-INF/services/geotrellis.layers.CollectionLayerReaderProvider
@@ -1,0 +1,1 @@
+geotrellis.layers.accumulo.AccumuloCollectionLayerProvider
