+++ layers-accumulo/src/main/scala/geotrellis/layers/accumulo/AccumuloCollectionLayerReader.scala
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package geotrellis.spark.io.accumulo
+package geotrellis.layers.accumulo
 
-import geotrellis.layers.{ContextCollection, LayerId}
 import geotrellis.tiling._
-import geotrellis.spark._
-import geotrellis.spark.io._
-import geotrellis.layers.io.avro._
+import geotrellis.layers._
+import geotrellis.layers.avro._
 import geotrellis.util._
+
 import org.apache.accumulo.core.data.{Range => AccumuloRange}
 import org.apache.hadoop.io.Text
+
 import spray.json._
 
 import scala.reflect._
